+++ PMS-PropertyHapa.Staff/Controllers/AccountTypeController.cs
@@ -2,21 +2,30 @@
 using PMS_PropertyHapa.Models.DTO;
 using PMS_PropertyHapa.Models.Entities;
 using PMS_PropertyHapa.Staff.Services.IServices;
+using static PMS_PropertyHapa.Shared.Enum.SD;
 
 namespace PMS_PropertyHapa.Staff.Controllers
 {
     public class AccountTypeController : Controller
     {
         private readonly IAuthService _authService;
-
-        public AccountTypeController(IAuthService authService)
+        private readonly IPermissionService _permissionService;
+        public AccountTypeController(IAuthService authService, IPermissionService permissionService)
         {
             _authService = authService;
+            _permissionService = permissionService;
         }
 
 
         public async Task<IActionResult> Index()
         {
+            var userId = Request?.Cookies["userId"]?.ToString();
+
+            bool hasAccess = await _permissionService.HasAccess(userId, (int)UserPermissions.ViewAccountType);
+            if (!hasAccess)
+            {
+                return Unauthorized();
+            }
             return View();
         }
 
@@ -37,6 +46,13 @@ namespace PMS_PropertyHapa.Staff.Controllers
         [HttpPost]
         public async Task<IActionResult> SaveAccountType([FromBody] AccountType accountType)
         {
+            var userId = Request?.Cookies["userId"]?.ToString();
+
+            bool hasAccess = await _permissionService.HasAccess(userId, (int)UserPermissions.AddAccountType);
+            if (!hasAccess)
+            {
+                return Unauthorized();
+            }
             if (accountType == null)
             {
                 return Json(new { success = false, message = "Received data is null." });
@@ -49,6 +65,13 @@ namespace PMS_PropertyHapa.Staff.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteAccountType(int id)
         {
+            var userId = Request?.Cookies["userId"]?.ToString();
+
+            bool hasAccess = await _permissionService.HasAccess(userId, (int)UserPermissions.AddAccountType);
+            if (!hasAccess)
+            {
+                return Unauthorized();
+            }
             var response = await _authService.DeleteAccountTypeAsync(id);
             if (!response.IsSuccess)
             {
